Allow modal to generate and show math problems
Currently this is just a WIP. A new function has been added that
generates an array of 15 arrays which stores the two numbers and
answer for a math problem. The jQuery event will call this function
and modify the contents of the modal to show the problem and answers.
The modal has been modified to receive this new data.