Add timer and grading functionality
A setTimeout call was added to the modal along with a new function
which alters the modal when time is up. The title changes (Thought
in the future the title should remain while a "time is up" textbar
should appear), the input boxes become disabled, any problems that
were not answered or answered incorrectly will have their background
turned to a red color.