Change test to show one problem at a time
Several changes needed to be made.
- All links to #testingModal were changed to #problemModal. Links
now include a .choice class for jQuery selectors.
- Added the problemModal, the main component for this change.
- The JS scripts now include a set of global variables to store
testing data. This will be changed very soon in the future but was
the easiest solution atm.
- Event listener for .choice click. This function will extract data
from the button the user clicked on, start the timer, and open the
problemModal.
- Event listener for #next button click. This will capture the given
answer, increment the index, and determine if the modal needs to
be updated with the next question or open the final modal.
- Event listener to see if an enter was pushed on the answer input
box. If so, trigger a button click on #next.
- Event listener triggered with the problem modal is shown.
- Updates to the testing modal so it correctly gets data.